[core] split geometry.h
