+++ src/xci/core/geometry/Vec2.h
@@ -1,17 +1,18 @@
-// geometry.h created on 2018-03-04 as part of xcikit project
+// Vec2.h created on 2018-03-04 as part of xcikit project
 // https://github.com/rbrich/xcikit
 //
 // Copyright 2018–2023 Radek Brich
 // Licensed under the Apache License, Version 2.0 (see LICENSE file)
 
-#ifndef XCI_CORE_GEOMETRY_H
-#define XCI_CORE_GEOMETRY_H
+#ifndef XCI_CORE_GEOMETRY_VEC2_H
+#define XCI_CORE_GEOMETRY_VEC2_H
 
 #include <cmath>
 #include <algorithm>
 #include <ostream>
 
 namespace xci::core {
+
 
 // Return value as POD numeric type.
 template<typename T, typename std::enable_if_t<std::is_arithmetic<T>::value, int> = 0>
@@ -79,7 +80,7 @@ struct Vec2 {
         return *this;
     }
 
-    operator bool() const noexcept {
+    explicit operator bool() const noexcept {
         return x != T{} || y != T{};
     }
 
@@ -154,95 +155,6 @@ using Vec2u = Vec2<uint32_t>;
 using Vec2f = Vec2<float>;
 
 
-template <typename T>
-struct Rect {
-    Rect() : x(0), y(0), w(0), h(0) {}
-    Rect(T x, T y, T w, T h) : x(x), y(y), w(w), h(h) {}
-    Rect(const Vec2<T>& pos, const Vec2<T>& size) : x(pos.x), y(pos.y), w(size.x), h(size.y) {}
-
-    bool contains(const Vec2<T>& point) const {
-        return point.x >= x && point.x <= x + w &&
-               point.y >= y && point.y <= y + h;
-    }
-    //bool intersects(const Rect<T>& rectangle) const;
-
-    Rect<T> union_(const Rect<T>& other) const {
-        auto l = std::min(x, other.x);
-        auto t = std::min(y, other.y);
-        auto r = std::max(right(), other.right());
-        auto b = std::max(bottom(), other.bottom());
-        return {l, t, r-l, b-t};
-    }
-
-    Rect<T> intersection(const Rect<T>& other) const {
-        auto l = std::max(x, other.x);
-        auto t = std::max(y, other.y);
-        auto r = std::min(right(), other.right());
-        auto b = std::min(bottom(), other.bottom());
-        return {l, t, r-l, b-t};
-    }
-
-    Rect<T> enlarged(T radius) const {
-        return {x - radius, y - radius, w + 2*radius, h + 2*radius};
-    }
-
-    Rect<T> moved(const Vec2<T>& offset) const {
-        return {x + offset.x, y + offset.y, w, h };
-    }
-
-    // Extend Rect to contain `other`
-    void extend(const Rect<T>& other) {
-        *this = union_(other);
-    }
-
-    void crop(const Rect<T>& other) {
-        *this = intersection(other);
-    }
-
-    // Enlarge Rect to all sides by `radius`
-    void enlarge(T radius) {
-        x -= radius;
-        y -= radius;
-        w += 2 * radius;
-        h += 2 * radius;
-    }
-
-    // Enlarge Rect to all sides by `radius`
-    void enlarge(const Vec2<T>& radius) {
-        x -= radius.x;
-        y -= radius.y;
-        w += 2 * radius.x;
-        h += 2 * radius.y;
-    }
-
-    inline T left() const { return x; }
-    inline T top() const { return y; }
-    inline T right() const { return x + w; }
-    inline T bottom() const { return y + h; }
-
-    inline Vec2<T> top_left() const { return {x, y}; }
-    inline Vec2<T> size() const { return {w, h}; }
-    inline Vec2<T> center() const { return {x + w/2, y + h/2}; }
-
-public:
-    T x;  // left
-    T y;  // top
-    T w;  // width
-    T h;  // height
-};
-
-template <typename T>
-std::ostream& operator <<(std::ostream& s, Rect<T> r) {
-    return s << "{" << r.x << ", " << r.y << ", "
-                    << r.w << ", " << r.h << "}";
-}
-
-using Rect_i = Rect<int32_t>;
-using Rect_u = Rect<uint32_t>;
-using Rect_f = Rect<float>;
-
-
-
 // Compute intersection between a line and a circle
 // Returns first intersection of line leading from `origin` in `direction`
 // with circle at `center` with `radius`, or INFINITY if there is no such
@@ -263,7 +175,7 @@ float line_circle_intersection(const Vec2<T> &origin,
         return INFINITY;
 
     discriminant = std::sqrt(discriminant);
-    float t = (-b - discriminant) / (2*a);
+    const float t = (-b - discriminant) / (2*a);
     if (t >= 0)
         return t;
     else
