+++ src/xci/text/Layout.h
@@ -14,7 +14,7 @@
 #include <xci/graphics/View.h>
 #include <xci/graphics/Color.h>
 #include <xci/graphics/shape/Rectangle.h>
-#include <xci/core/geometry.h>
+#include <xci/core/geometry/Vec2.h>
 #include <xci/core/container/ChunkedStack.h>
 
 #include <string>
